Smoothed IRC protocol, fixed frequent stuckness, continues to be connected.
The bot doesn't rape the IRC protocol anymore, which means the server
will be much more forgiving towards our terse implementation. Fixed the
frequent deadlock caused by the blocking "oef?". The bot now continues
to stay connected instead of hopping on and off the server, which caused
the connections to be throttled.